backend/model/shap_explainer.py
"""
SHAP Explainability Module - Feature Importance Analysis

Provides SHAP-based explanations for ensemble predictions.
Shows which features most influenced each prediction.
"""

import numpy as np
import pickle
import os
from shap import TreeExplainer, force_plot, summary_plot
import matplotlib
matplotlib.use('Agg')  # Non-interactive backend
import matplotlib.pyplot as plt

from backend.utils.helpers import setup_logger, log_step
from backend.preprocessing.feature_engineering import prepare_feature_vector

logger = setup_logger(__name__)

ENSEMBLE_MODELS_PATH = 'backend/model/saved/ensemble/'


class SHAPExplainer:
    """SHAP-based explainability for ensemble predictions."""
    
    def __init__(self):
        """Initialize SHAP explainers for RF and XGBoost."""
        self.rf_model = None
        self.xgb_model = None
        self.rf_explainer = None
        self.xgb_explainer = None
        self.scaler = None
        
        self.feature_names = [
            'NDVI Mean', 'NDVI Trend', 'NDVI Variance',
            'Rainfall Deviation %', 'Temperature Anomaly °C',
            'Soil Moisture Index', 'Soil Type (1-5)', 'Pest Frequency'
        ]
        
        self.load_models()
    
    def load_models(self):
        """Load RF and XGBoost models, create SHAP explainers."""
        try:
            log_step("Loading SHAP Explainers", "in_progress")
            
            # Load models
            with open(f'{ENSEMBLE_MODELS_PATH}rf_model.pkl', 'rb') as f:
                self.rf_model = pickle.load(f)
            logger.info("✓ Loaded RF model for SHAP")
            
            with open(f'{ENSEMBLE_MODELS_PATH}xgb_model.pkl', 'rb') as f:
                self.xgb_model = pickle.load(f)
            logger.info("✓ Loaded XGBoost model for SHAP")
            
            # Load scaler
            with open(f'{ENSEMBLE_MODELS_PATH}scaler.pkl', 'rb') as f:
                self.scaler = pickle.load(f)
            logger.info("✓ Loaded feature scaler")
            
            # Create SHAP explainers
            self.rf_explainer = TreeExplainer(self.rf_model)
            logger.info("✓ Created RF SHAP explainer")
            
            self.xgb_explainer = TreeExplainer(self.xgb_model)
            logger.info("✓ Created XGBoost SHAP explainer")
            
            log_step("Loading SHAP Explainers", "success")
            
        except Exception as e:
            logger.error(f"Failed to load SHAP explainers: {e}")
            raise
    
    def explain_prediction(self, state, district, crop, season):
        """
        Generate explanation for prediction using feature importance.
        
        Returns:
            {
                'feature_importance': [
                    {'feature': 'NDVI Mean', 'value': 0.687, 'contribution': 0.12, 'direction': 'increases_risk'},
                    ...
                ],
                'rf_top_features': [...],
                'xgb_top_features': [...],
                'prediction_logic': 'Natural language explanation'
            }
        """
        try:
            log_step("SHAP Explanation Generation", "in_progress")
            
            # Prepare features
            norm_features, raw_features = prepare_feature_vector(state, district, crop, season)
            
            # Create feature vector
            feature_vector = np.array([
                norm_features['ndvi_mean'],
                norm_features['ndvi_trend'],
                norm_features['ndvi_variance'],
                norm_features['rainfall_deviation'],
                norm_features['temperature_anomaly'],
                norm_features['soil_moisture_index'],
                norm_features['soil_type_encoded'],
                norm_features['pest_frequency']
            ]).reshape(1, -1)
            
            # Scale features
            if self.scaler is not None:
                feature_vector_scaled = self.scaler.transform(feature_vector)
            else:
                feature_vector_scaled = feature_vector
            
            # Get model feature importances (deterministic, no SHAP complexity)
            rf_importance_dict = dict(zip(self.feature_names, self.rf_model.feature_importances_))
            xgb_importance_dict = dict(zip(self.feature_names, self.xgb_model.feature_importances_))
            
            # Average importances
            avg_importance = {}
            for fname in self.feature_names:
                avg_importance[fname] = (rf_importance_dict[fname] + xgb_importance_dict[fname]) / 2
            
            # Create feature importance list with directional analysis
            feature_importance = []
            for fname, raw_val in zip(self.feature_names, feature_vector_scaled[0]):
                importance = avg_importance[fname]
                
                # Determine direction based on feature value vs risk relationship
                if 'ndvi' in fname.lower() or 'moisture' in fname.lower():
                    # These decrease risk when high
                    direction = "decreases_risk" if raw_val > 0 else "increases_risk"
                elif 'rainfall' in fname.lower() or 'temperature' in fname.lower() or 'pest' in fname.lower():
                    # These increase risk when anomalous
                    direction = "increases_risk" if abs(raw_val) > 0 else "decreases_risk"
                else:
                    direction = "increases_risk" if raw_val > 0.5 else "decreases_risk"
                
                feature_importance.append({
                    'feature': fname,
                    'raw_value': float(raw_val),
                    'contribution': importance,
                    'direction': direction,
                    'impact': 'High' if importance > np.mean(list(avg_importance.values())) else 'Low'
                })
            
            # Sort by contribution
            feature_importance.sort(key=lambda x: x['contribution'], reverse=True)
            
            # Get top features for each model
            rf_top = sorted(rf_importance_dict.items(), key=lambda x: x[1], reverse=True)[:3]
            xgb_top = sorted(xgb_importance_dict.items(), key=lambda x: x[1], reverse=True)[:3]
            
            # Generate natural language explanation
            top_positive = [f for f in feature_importance if f['direction'] == 'increases_risk'][:2]
            top_negative = [f for f in feature_importance if f['direction'] == 'decreases_risk'][:2]
            
            explanation_parts = []
            for f in top_positive[:1]:
                if 'NDVI' in f['feature']:
                    explanation_parts.append(f"Low vegetation health is the primary risk driver")
                elif 'Rainfall' in f['feature']:
                    explanation_parts.append(f"Rainfall variability increases failure risk")
                elif 'Soil' in f['feature'] and 'Index' in f['feature']:
                    explanation_parts.append(f"Poor soil conditions elevate risk")
                elif 'Pest' in f['feature']:
                    explanation_parts.append(f"High pest activity threatens the crop")
                elif 'Temperature' in f['feature']:
                    explanation_parts.append(f"Temperature stress stresses the crop")
            
            for f in top_negative[:1]:
                if 'NDVI' in f['feature']:
                    explanation_parts.append(f"Good vegetation health mitigates risk")
                elif 'Moisture' in f['feature']:
                    explanation_parts.append(f"Adequate soil moisture provides resilience")
            
            prediction_logic = "; ".join(explanation_parts) if explanation_parts else \
                "Prediction based on balanced feature interactions"
            
            result = {
                'feature_importance': feature_importance,
                'rf_top_features': [{'feature': name, 'importance': float(val)} for name, val in rf_top],
                'xgb_top_features': [{'feature': name, 'importance': float(val)} for name, val in xgb_top],
                'prediction_logic': prediction_logic,
                'raw_features': raw_features
            }
            
            log_step("SHAP Explanation Generation", "success")
            
            return result
            
        except Exception as e:
            logger.error(f"SHAP explanation failed: {e}")
            raise


# Singleton instance
_shap_explainer = None


def get_shap_explainer():
    """Get or create singleton SHAP explainer."""
    global _shap_explainer
    if _shap_explainer is None:
        _shap_explainer = SHAPExplainer()
    return _shap_explainer


def explain_ensemble_prediction(state, district, crop, season):
    """
    Unified SHAP explanation API.
    """
    explainer = get_shap_explainer()
    return explainer.explain_prediction(state, district, crop, season)
