requirements.txt
# Backend Dependencies
Flask>=3.0.0
Flask-CORS>=4.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
joblib>=1.3.0
Werkzeug>=3.0.0
reportlab>=4.0.0

# Data Processing
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
